refactor: ReservationRepository 변경사항에 따른 수정
